+++ src/popup/component/TransactionModal/index.js
@@ -130,8 +130,10 @@ export const TransactionModal = ({
                     </div>
                   </div>
                   {modalType === TransactionModalType.speedUp && (
-                    <div className={styles.changeCls} onClick={onClickAdvance}>
-                      {i18n.t("change")}
+                    <div className={styles.changeWrapper}>
+                      <div className={styles.changeCls} onClick={onClickAdvance}>
+                        {i18n.t("change")} 
+                      </div>
                     </div>
                   )}
                 </div>